refactor: rename app header
